requirements.txt
joblib==1.4.2
matplotlib==3.9.2
numpy==2.1.1
pandas==2.2.2
Pillow==10.4.0
plotly==5.24.0
seaborn==0.13.2
streamlit==1.38.0
